#14 created status bar from bootstrap, add logic to calculate the width.
